Remove impl_block module
